Refactor xgettext to make it testable by fuzzers
